Refactor EWallet Service, add Tokenization flow, update examples
